add single layout with new app provider
